requirements.txt
typer==0.12.5
rich==13.9.4
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
requests==2.32.3
rapidfuzz==3.10.0
vaderSentiment==3.3.2
streamlit==1.39.0

